Exceptions added for delete, update and create Few travel transaction disputes were solved

+++ spaceshipProject/src/main/java/com/example/demo/controllers/spaceshipController.java
@@ -2,16 +2,13 @@ package com.example.demo.controllers;
 
 import org.springframework.web.bind.annotation.RestController;
 
+import com.example.demo.model.Location;
 import com.example.demo.model.Spaceship;
+import com.example.demo.repositories.locationRepository;
 import com.example.demo.repositories.spaceshipRepository;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.mongodb.core.ExecutableUpdateOperation;
-import org.springframework.data.mongodb.core.MongoOperations;
-import org.springframework.data.mongodb.core.query.Update;
-import org.springframework.data.mongodb.core.query.Query;
-import org.springframework.data.mongodb.core.MongoTemplate;
-import org.springframework.data.mongodb.core.query.Criteria;
+
 import org.springframework.web.bind.annotation.GetMapping;
 
 import org.springframework.web.bind.annotation.PathVariable;
@@ -20,6 +17,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
+import java.util.ArrayList;
 import java.util.List;
 
 
@@ -30,6 +28,11 @@ public class spaceshipController
 	
 	@Autowired
 	public spaceshipRepository spaceRepo;
+	
+	@Autowired
+	public locationRepository locationRepo;
+	
+
 	
 	@GetMapping (value = "/all")
 	public List<Spaceship> getallSpaceships()
@@ -42,15 +45,64 @@ public class spaceshipController
 	@PostMapping (value = "/create")
 	public String createSpaceship(@RequestBody Spaceship spaceship)
 	{
+		List<Location> location = new ArrayList<>(locationRepo.findAll());
+		List<Spaceship> space = new ArrayList<>(spaceRepo.findAll());
 		
-		Spaceship insertedship = spaceRepo.insert(spaceship);
-		return "Spaceship created "+insertedship.getName();
+		int count = 0;
+		
+		for(int i=0;i<space.size();i++)
+		{
+			
+			Spaceship s = space.get(i);
+			
+			if(s.getId() == spaceship.getId())
+			{
+				return "Spaceship exists already";
+			}
+		}
+		for(int i=0;i<location.size();i++)
+		{
+			 Location l = location.get(i);
+			 
+			if(spaceship.getlocation_Id() == l.getId())
+			{
+				spaceship.setLocation(l.getCityName()+" "+l.getPlanetName());
+				
+				Spaceship insertedship = spaceRepo.insert(spaceship);
+				return "Spaceship created "+insertedship.getName();
+			}
+			
+			else
+			{
+				return "Location was not found";
+			}
+		}
+		return "";
+		
+		
 	}
 	
 	@RequestMapping (value = "/delete/{Id}", method = RequestMethod.POST)
 	public String deleteRecord(@PathVariable("Id") long Id) 
 	{
+		int count =0;
+		List<Spaceship> space = new ArrayList<>(spaceRepo.findAll());
+
+		for(int i=0;i<space.size();i++)
+		{
+			
+			Spaceship s = space.get(i);
+			
+			if(s.getId() == Id)
+			{
+				count++;
+			}
+		}
 		
+		if(count == 0)
+		{
+			return "Cannot delete :( since Spaceship ID is not found";
+		}
 		 this.spaceRepo.deleteById(Id);
 	    return "Successfully deleted";
 	}
@@ -58,6 +110,25 @@ public class spaceshipController
 	@RequestMapping (value = "/update/{id}", method = RequestMethod.POST)
 	public String updateRecord(@RequestBody Spaceship spaceship) 
 	{
+		
+		int count = 0;
+		List<Spaceship> space = new ArrayList<>(spaceRepo.findAll());
+
+		for(int i=0;i<space.size();i++)
+		{
+			
+			Spaceship s = space.get(i);
+			
+			if(s.getId() == spaceship.getId())
+			{
+				count++;
+			}
+		}
+		
+		if(count == 0)
+		{
+			return "Cannot update :( since Spaceship ID is not found";
+		}
 		
 		String status = spaceship.getStatus();
 		spaceship.setStatus(status);
